Transacciones financiera rest

+++ .NET_Negocio/LogicaFinanciera/Movimiento.cs
@@ -14,12 +14,12 @@ namespace LogicaFinanciera
     
     public partial class Movimiento
     {
-        public decimal id_movimiento { get; set; }
+        public int id_movimiento { get; set; }
         public Nullable<decimal> num_aprovacion { get; set; }
-        public decimal valor { get; set; }
+        public int valor { get; set; }
         public System.DateTime fecha { get; set; }
         public bool estado { get; set; }
-        public decimal id_usuario { get; set; }
+        public int id_usuario { get; set; }
     
         public virtual Usuario Usuario { get; set; }
     }
