+++ .NET_Negocio/LogicaFinanciera/Usuario.cs
@@ -20,12 +20,12 @@ namespace LogicaFinanciera
             this.Movimiento = new HashSet<Movimiento>();
         }
     
-        public decimal id_usuario { get; set; }
+        public int id_usuario { get; set; }
         public string nombre { get; set; }
         public string password { get; set; }
         public string tipo_documento { get; set; }
-        public decimal numero_documento { get; set; }
-        public decimal saldo { get; set; }
+        public int numero_documento { get; set; }
+        public int saldo { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Movimiento> Movimiento { get; set; }
